moved test, added to conf, fixed test
